Vanilla-ize some lodash calls to speed things up
Turns out lodash is pretty slow. This commit ends up speeding up
redrawing the swatch by about 1.5x. Combined with the other
optimizations, it comes out to around 2x, which is not enough to where
rerendering is buttery smooth, but enough to where it's at least a lot
more usable.